+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.12.2] - 2020-02-05
+
 -   更新依赖包 minor 和 patch 版本号
 
 **koot**